adding license and cleanup
